Terminaison des tests et avancement de l'application

+++ back/src/main/java/com/openclassrooms/mddapi/repository/ArticleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends CrudRepository<DBArticle, Integer> {
+public interface ArticleRepository extends CrudRepository<DBArticle, Long> {
 }
